refactor(desktop): the errand flight is a hook
